+++ packages/getters/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @penumbra-zone/getters
 
+## 26.0.0
+
+### Patch Changes
+
+- Updated dependencies [93f1d05]
+  - @penumbra-zone/protobuf@10.0.0
+  - @penumbra-zone/bech32m@16.0.0
+
 ## 25.0.1
 
 ### Patch Changes